[universal-stdout-logs] Use tail -F to follow rotated log files
I appreciate this mod for keeping an eye on nginx logs and such. The lsio/swag rotates those log files daily, and unfortunately that causes this mod to stop working. I propose using tail's `-F` (a shortcut for `--follow=name --retry`) instead of `-f` so that this mod will continue tailing the newly-rotated log file.