calibre-web: dtrpg-metadata Add DMs Guild and fix cover downloads
DriveThruRPG contains multiple stores, one of which is the DMs Guild
for Dungeons and Dragons. Items that are only for sale on the Guild
don't show up on DTRPG searches, and vice versa.

A lot of the images end up being redirected to an eventual file, but
Calibre-web doesn't follow redirects when saving the cover. Instead, we
follow them up front and return the eventual URL.